feat(mcp): discover needs-auth opens the shared Connect login modal (#459)
Instead of only expanding the row, a discover "needs authorization" now opens the
SAME login modal as the chat auth-required card. Extracted the chat card modal
into a shared _components/mcp/McpConnectModal (wraps McpAuthSection); the chat
card and the Control Center discover prompt both use it, so the login experience
is identical. tsc + i18n + lint green.
