+++ web-ui/app/_components/chat/McpAuthRequiredCard.tsx
@@ -1,10 +1,10 @@
 'use client';
 
-import { useEffect, useState } from 'react';
+import { useState } from 'react';
 import { useTranslations } from 'next-intl';
 
 import { Button } from '../ui/Button';
-import { McpAuthSection } from '../mcp/McpAuthSection';
+import { McpConnectModal } from '../mcp/McpConnectModal';
 
 /**
  * Epic #459 W9 — in-chat "this MCP server needs authorization" card.
@@ -81,68 +81,6 @@ export function parseMcpAuthRequired(content: string): ParseResult {
   };
 }
 
-function ConnectModal({
-  auth,
-  onClose,
-}: {
-  auth: ParsedMcpAuthRequired;
-  onClose: () => void;
-}): React.ReactElement {
-  const t = useTranslations('chat');
-
-  useEffect(() => {
-    const onKey = (e: globalThis.KeyboardEvent): void => {
-      if (e.key === 'Escape') {
-        e.preventDefault();
-        onClose();
-      }
-    };
-    window.addEventListener('keydown', onKey);
-    return () => {
-      window.removeEventListener('keydown', onKey);
-    };
-  }, [onClose]);
-
-  return (
-    <div
-      role="dialog"
-      aria-modal="true"
-      aria-label={t('mcpAuth.modalTitle', { server: auth.server })}
-      className="fixed inset-0 z-50 flex items-center justify-center bg-[color:var(--bg-modal-overlay)] p-4"
-      onClick={(e) => {
-        if (e.target === e.currentTarget) onClose();
-      }}
-    >
-      <div className="w-full max-w-md rounded-lg border border-[color:var(--border)] bg-[color:var(--bg-elevated)] p-4 shadow-lg">
-        <div className="flex items-start justify-between gap-3">
-          <h2 className="text-base font-semibold text-[color:var(--fg-strong)]">
-            {t('mcpAuth.modalTitle', { server: auth.server })}
-          </h2>
-          <button
-            type="button"
-            onClick={onClose}
-            aria-label={t('mcpAuth.close')}
-            className="text-[color:var(--fg-muted)] hover:text-[color:var(--fg-strong)]"
-          >
-            ✕
-          </button>
-        </div>
-        <p className="mt-2 text-sm text-[color:var(--fg-muted)]">
-          {t('mcpAuth.modalBody', { server: auth.server })}
-        </p>
-        <div className="mt-3">
-          <McpAuthSection serverId={auth.serverId} />
-        </div>
-        <div className="mt-4 flex justify-end">
-          <Button variant="secondary" size="sm" onClick={onClose}>
-            {t('mcpAuth.close')}
-          </Button>
-        </div>
-      </div>
-    </div>
-  );
-}
-
 export function McpAuthRequiredCard({
   auth,
 }: {
@@ -165,7 +103,13 @@ export function McpAuthRequiredCard({
           {t('mcpAuth.connect', { server: auth.server })}
         </Button>
       </div>
-      {open ? <ConnectModal auth={auth} onClose={() => setOpen(false)} /> : null}
+      {open ? (
+        <McpConnectModal
+          serverId={auth.serverId}
+          serverName={auth.server}
+          onClose={() => setOpen(false)}
+        />
+      ) : null}
     </div>
   );
 }
